# openapi/el-trigal-api.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "El Trigal API",
    "version": "1.0.0",
    "description": "API REST independiente para el website Astro y el panel Angular."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "paths": {
...
    "/public/categories": {
      "get": {
        "tags": ["public/catalog"],
        "operationId": "listPublicCategories",
        "parameters": [
          { "$ref": "#/components/parameters/locale" },
          { "$ref": "#/components/parameters/page" },
          { "$ref": "#/components/parameters/limit" },
          { "$ref": "#/components/parameters/search" },
          { "$ref": "#/components/parameters/featured" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/CategoryList" },
          "400": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
    "/public/categories/{slug}": {
      "get": {
        "tags": ["public/catalog"],
        "operationId": "getPublicCategoryBySlug",
        "parameters": [
          { "$ref": "#/components/parameters/slug" },
          { "$ref": "#/components/parameters/locale" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/CategoryItem" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/public/products": {
      "get": {
        "tags": ["public/catalog"],
        "operationId": "listPublicProducts",
        "parameters": [
          { "$ref": "#/components/parameters/locale" },
          { "$ref": "#/components/parameters/page" },
          { "$ref": "#/components/parameters/limit" },
          { "$ref": "#/components/parameters/search" },
          { "$ref": "#/components/parameters/featured" },
          { "$ref": "#/components/parameters/bestSeller" },
          { "$ref": "#/components/parameters/newProduct" },
          { "$ref": "#/components/parameters/availability" },
          { "$ref": "#/components/parameters/productSort" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/ProductList" },
          "400": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
    "/public/products/{slug}": {
      "get": {
        "tags": ["public/catalog"],
        "operationId": "getPublicProductBySlug",
        "parameters": [
          { "$ref": "#/components/parameters/slug" },
          { "$ref": "#/components/parameters/locale" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/ProductItem" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/public/products/{slug}/related": {
      "get": {
        "tags": ["public/catalog"],
        "operationId": "listRelatedProducts",
        "parameters": [
          { "$ref": "#/components/parameters/slug" },
          { "$ref": "#/components/parameters/locale" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/ProductArray" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/auth/login": {
      "post": {
        "tags": ["auth"],
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } } }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/AuthSuccess" },
          "401": { "$ref": "#/components/responses/InvalidCredentials" },
          "403": { "$ref": "#/components/responses/AccountLocked" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["auth"],
        "operationId": "refresh",
        "description": "Usa cookie HttpOnly el_trigal_refresh y rota el refresh token.",
        "responses": {
          "200": { "$ref": "#/components/responses/AuthSuccess" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/auth/forgot-password": {
      "post": {
        "tags": ["auth"],
        "operationId": "forgotPassword",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ForgotPasswordRequest" } } }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/AcceptedMessage" },
          "400": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "tags": ["auth"],
        "operationId": "resetPassword",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResetPasswordRequest" } } }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/AcceptedMessage" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["auth"],
        "operationId": "getMe",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/UserItem" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/auth/change-password": {
      "post": {
        "tags": ["auth"],
        "operationId": "changePassword",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ChangePasswordRequest" } } }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/AcceptedMessage" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/admin/products": {
      "get": {
        "tags": ["admin/products"],
        "operationId": "listAdminProducts",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/page" },
          { "$ref": "#/components/parameters/limit" },
          { "$ref": "#/components/parameters/status" },
          { "$ref": "#/components/parameters/search" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/ProductList" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "post": {
        "tags": ["admin/products"],
        "operationId": "createProduct",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProductInput" } } }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/ProductItem" },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/admin/products/{id}": {
      "get": {
        "tags": ["admin/products"],
        "operationId": "getAdminProduct",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/id" }],
        "responses": { "200": { "$ref": "#/components/responses/ProductItem" }, "404": { "$ref": "#/components/responses/NotFound" } }
      },
      "patch": {
        "tags": ["admin/products"],
        "operationId": "updateProduct",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/id" }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProductPatch" } } }
        },
        "responses": { "200": { "$ref": "#/components/responses/ProductItem" }, "400": { "$ref": "#/components/responses/ValidationError" }, "404": { "$ref": "#/components/responses/NotFound" } }
      },
      "delete": {
        "tags": ["admin/products"],
        "operationId": "archiveProductByDelete",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/id" }],
        "responses": { "204": { "description": "Producto archivado sin cuerpo." }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
...
    "/admin/categories": {
      "get": {
        "tags": ["admin/categories"],
        "operationId": "listAdminCategories",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/page" }, { "$ref": "#/components/parameters/limit" }],
        "responses": { "200": { "$ref": "#/components/responses/CategoryList" } }
      },
      "post": {
        "tags": ["admin/categories"],
        "operationId": "createCategory",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CategoryInput" } } } },
        "responses": { "201": { "$ref": "#/components/responses/CategoryItem" }, "400": { "$ref": "#/components/responses/ValidationError" }, "409": { "$ref": "#/components/responses/Conflict" } }
      }
    },
    "/admin/categories/{id}": {
      "get": {
        "tags": ["admin/categories"],
        "operationId": "getAdminCategory",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/id" }],
        "responses": { "200": { "$ref": "#/components/responses/CategoryItem" }, "404": { "$ref": "#/components/responses/NotFound" } }
      },
      "patch": {
        "tags": ["admin/categories"],
        "operationId": "updateCategory",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/id" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CategoryPatch" } } } },
        "responses": { "200": { "$ref": "#/components/responses/CategoryItem" }, "400": { "$ref": "#/components/responses/ValidationError" } }
      },
      "delete": {
        "tags": ["admin/categories"],
        "operationId": "archiveCategoryByDelete",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/id" }],
        "responses": { "204": { "description": "Categoria archivada sin cuerpo." }, "409": { "$ref": "#/components/responses/CategoryNotEmpty" } }
      }
    },
    "/admin/categories/reorder": {
      "patch": {
        "tags": ["admin/categories"],
        "operationId": "reorderCategories",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReorderCategoriesRequest" } } } },
        "responses": { "200": { "$ref": "#/components/responses/AcceptedMessage" } }
      }
    },
...
    "/admin/users": {
      "get": {
        "tags": ["admin/users"],
        "operationId": "listUsers",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "$ref": "#/components/responses/UserList" } }
      },
      "post": {
        "tags": ["admin/users"],
        "operationId": "createUser",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateUserRequest" } } } },
        "responses": { "201": { "$ref": "#/components/responses/UserItem" }, "409": { "$ref": "#/components/responses/Conflict" } }
      }
    },
    "/admin/users/{id}": {
      "patch": {
        "tags": ["admin/users"],
        "operationId": "updateUser",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/id" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateUserRequest" } } } },
        "responses": { "200": { "$ref": "#/components/responses/UserItem" } }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "parameters": {
      "id": { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "pattern": "^[a-fA-F0-9]{24}$" } },
      "slug": { "name": "slug", "in": "path", "required": true, "schema": { "type": "string", "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$" } },
      "locale": { "name": "locale", "in": "query", "required": true, "schema": { "$ref": "#/components/schemas/Locale" } },
      "page": { "name": "page", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 10000, "default": 1 } },
      "limit": { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 24 } },
      "search": { "name": "search", "in": "query", "required": false, "schema": { "type": "string", "minLength": 2, "maxLength": 80 } },
      "featured": { "name": "featured", "in": "query", "required": false, "schema": { "type": "boolean" } },
      "bestSeller": { "name": "bestSeller", "in": "query", "required": false, "schema": { "type": "boolean" } },
      "newProduct": { "name": "newProduct", "in": "query", "required": false, "schema": { "type": "boolean" } },
      "availability": { "name": "availability", "in": "query", "required": false, "schema": { "$ref": "#/components/schemas/AvailabilityStatus" } },
      "productSort": { "name": "sort", "in": "query", "required": false, "schema": { "type": "string", "enum": ["sortOrder", "name", "price-asc", "price-desc", "newest"], "default": "sortOrder" } },
      "status": { "name": "status", "in": "query", "required": false, "schema": { "$ref": "#/components/schemas/Status" } }
    },
...
    "responses": {
      "ValidationError": { "description": "Datos invalidos.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
      "Unauthorized": { "description": "No autenticado.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
      "Forbidden": { "description": "Sin permiso.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
      "InvalidCredentials": { "description": "Credenciales invalidas.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
      "AccountLocked": { "description": "Cuenta bloqueada temporalmente.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
      "NotFound": { "description": "No encontrado.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
      "Conflict": { "description": "Conflicto.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
      "RateLimited": { "description": "Rate limit.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
      "CategoryNotEmpty": { "description": "Categoria con productos.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
...
      "AcceptedMessage": { "description": "Mensaje aceptado.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MessageEnvelope" } } } },
      "AuthSuccess": { "description": "Token access y usuario.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthEnvelope" } } } },
      "UserItem": { "description": "Usuario.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserEnvelope" } } } },
      "UserList": { "description": "Usuarios.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserListEnvelope" } } } },
      "CategoryItem": { "description": "Categoria.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CategoryEnvelope" } } } },
      "CategoryList": { "description": "Categorias.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CategoryListEnvelope" } } } },
      "ProductItem": { "description": "Producto.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProductEnvelope" } } } },
      "ProductList": { "description": "Productos.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProductListEnvelope" } } } },
      "ProductArray": { "description": "Productos.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProductArrayEnvelope" } } } },
...
    },
    "schemas": {
      "Locale": { "type": "string", "enum": ["es", "en"] },
...
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "const": false },
          "error": {
            "type": "object",
            "properties": {
              "code": { "$ref": "#/components/schemas/ErrorCode" },
              "message": { "type": "string" },
              "details": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "field": { "type": "string" },
                    "rule": { "type": "string" },
                    "message": { "type": "string" }
                  },
                  "required": ["field", "rule", "message"],
                  "additionalProperties": false
                }
              }
            },
            "required": ["code", "message"],
            "additionalProperties": false
          },
          "meta": { "$ref": "#/components/schemas/Meta" }
        },
        "required": ["success", "error", "meta"],
        "additionalProperties": false
      },
...
      "CategoryInput": {
        "type": "object",
        "properties": {
          "parentId": { "type": ["string", "null"], "pattern": "^[a-fA-F0-9]{24}$" },
          "name": { "$ref": "#/components/schemas/LocalizedText" },
          "slug": { "$ref": "#/components/schemas/LocalizedText" },
          "description": { "anyOf": [{ "$ref": "#/components/schemas/LocalizedText" }, { "type": "null" }] },
          "shortDescription": { "anyOf": [{ "$ref": "#/components/schemas/LocalizedText" }, { "type": "null" }] },
          "imageId": { "type": ["string", "null"], "pattern": "^[a-fA-F0-9]{24}$" },
          "status": { "$ref": "#/components/schemas/Status" },
          "featured": { "type": "boolean" },
          "sortOrder": { "type": "integer", "minimum": 0, "maximum": 100000 },
          "seo": { "$ref": "#/components/schemas/SeoFields" }
        },
        "required": ["name", "slug", "featured", "sortOrder"],
        "additionalProperties": false
      },
      "CategoryPatch": { "$ref": "#/components/schemas/CategoryInput" },
...
      "ProductInput": {
        "type": "object",
        "properties": {
          "sku": { "type": ["string", "null"], "pattern": "^[A-Z0-9][A-Z0-9._-]{1,39}$" },
          "name": { "$ref": "#/components/schemas/LocalizedText" },
          "slug": { "$ref": "#/components/schemas/LocalizedText" },
          "shortDescription": { "$ref": "#/components/schemas/LocalizedText" },
          "description": { "$ref": "#/components/schemas/LocalizedText" },
          "primaryCategoryId": { "type": "string", "pattern": "^[a-fA-F0-9]{24}$" },
          "categoryIds": { "type": "array", "minItems": 1, "maxItems": 20, "uniqueItems": true, "items": { "type": "string", "pattern": "^[a-fA-F0-9]{24}$" } },
          "productType": { "$ref": "#/components/schemas/ProductType" },
          "basePriceCents": { "type": ["integer", "null"], "minimum": 0, "maximum": 100000000 },
          "compareAtPriceCents": { "type": ["integer", "null"], "minimum": 0, "maximum": 100000000 },
          "currency": { "type": "string", "const": "USD" },
          "priceLabel": { "anyOf": [{ "$ref": "#/components/schemas/LocalizedText" }, { "type": "null" }] },
          "variants": { "type": "array", "maxItems": 50, "items": { "$ref": "#/components/schemas/ProductVariant" } },
          "options": { "type": "array", "items": { "$ref": "#/components/schemas/ProductOption" } },
          "media": { "type": "array", "maxItems": 20, "items": { "$ref": "#/components/schemas/MediaAsset" } },
          "ingredients": { "anyOf": [{ "$ref": "#/components/schemas/LocalizedText" }, { "type": "null" }] },
          "allergens": { "type": "array", "uniqueItems": true, "items": { "$ref": "#/components/schemas/Allergen" } },
          "dietaryTags": { "type": "array", "maxItems": 30, "items": { "type": "string" } },
          "availability": { "$ref": "#/components/schemas/Availability" },
          "ordering": { "$ref": "#/components/schemas/OrderingOptions" },
          "preparation": { "anyOf": [{ "$ref": "#/components/schemas/Preparation" }, { "type": "null" }] },
          "tags": { "type": "array", "maxItems": 30, "items": { "type": "string" } },
          "featured": { "type": "boolean" },
          "bestSeller": { "type": "boolean" },
          "newProduct": { "type": "boolean" },
          "sortOrder": { "type": "integer", "minimum": 0, "maximum": 100000 },
          "seo": { "$ref": "#/components/schemas/SeoFields" },
          "status": { "$ref": "#/components/schemas/Status" }
        },
        "required": ["name", "slug", "shortDescription", "description", "primaryCategoryId", "categoryIds", "productType", "currency", "variants", "options", "media", "allergens", "dietaryTags", "availability", "ordering", "tags", "featured", "bestSeller", "newProduct", "sortOrder"],
        "additionalProperties": false
      },
      "ProductPatch": { "$ref": "#/components/schemas/ProductInput" },
...
      "LoginRequest": {
        "type": "object",
        "properties": { "email": { "type": "string", "format": "email", "maxLength": 254 }, "password": { "type": "string", "minLength": 8, "maxLength": 128 } },
        "required": ["email", "password"],
        "additionalProperties": false
      },
      "ForgotPasswordRequest": {
        "type": "object",
        "properties": { "email": { "type": "string", "format": "email", "maxLength": 254 } },
        "required": ["email"],
        "additionalProperties": false
      },
      "ResetPasswordRequest": {
        "type": "object",
        "properties": { "token": { "type": "string", "minLength": 32 }, "password": { "type": "string", "minLength": 12, "maxLength": 128 } },
        "required": ["token", "password"],
        "additionalProperties": false
      },
      "ChangePasswordRequest": {
        "type": "object",
        "properties": { "currentPassword": { "type": "string", "minLength": 8, "maxLength": 128 }, "newPassword": { "type": "string", "minLength": 12, "maxLength": 128 } },
        "required": ["currentPassword", "newPassword"],
        "additionalProperties": false
      },
...
      "CreateUserRequest": {
        "type": "object",
        "properties": { "email": { "type": "string", "format": "email" }, "role": { "type": "string", "enum": ["OWNER", "ADMIN", "EDITOR"] }, "password": { "type": "string", "minLength": 12, "maxLength": 128 } },
        "required": ["email", "role", "password"],
        "additionalProperties": false
      },
      "UpdateUserRequest": { "$ref": "#/components/schemas/CreateUserRequest" },
...
      "ReorderCategoriesRequest": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "maxItems": 200,
            "items": {
              "type": "object",
              "properties": { "id": { "type": "string", "pattern": "^[a-fA-F0-9]{24}$" }, "sortOrder": { "type": "integer", "minimum": 0, "maximum": 100000 } },
              "required": ["id", "sortOrder"],
              "additionalProperties": false
            }
          }
        },
        "required": ["items"],
        "additionalProperties": false
      },
...
      "CategoryListEnvelope": { "allOf": [{ "$ref": "#/components/schemas/SuccessEnvelope" }, { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/Category" } }, "meta": { "$ref": "#/components/schemas/PaginationMeta" } } }] },
...
    }
  }
}